gcloud/my-terraform-project/data-services/v1/ingestion/cloud_functions/requirements.txt
flask
google-cloud-storage
pandas
pyarrow